refactor: 특정 Entitiy만 다루는 Component 분리
