+++ src/refactoring/components/admin/coupon/CouponList.tsx
@@ -1,4 +1,4 @@
-import { Coupon } from '../../../types'
+import { Coupon } from '../../../../types'
 import { CouponBox } from './CouponBox'
 
 interface Props {
