+++ src/refactoring/components/admin/coupon/CreateCouponFormBox.tsx
@@ -1,4 +1,4 @@
-import { Coupon } from '../../../types'
+import { Coupon } from '../../../../types'
 
 interface Props {
   newCoupon: Coupon
